refactor(client-debugger-view): Move `ContainerHistoryLog` into `components` directory (#15328)
This module / component is really an implementation detail of the
visualizer, so it doesn't belong at the package root.
